Extract HealthZ handler to common class
