+++ test/IntegrationTests/Helpers/MockLogsCollector.cs
@@ -35,7 +35,7 @@ public class MockLogsCollector : IDisposable
     private readonly ITestOutputHelper _output;
     private readonly TestHttpListener _listener;
     private readonly BlockingCollection<global::OpenTelemetry.Proto.Logs.V1.LogRecord> _logs = new(100); // bounded to avoid memory leak
-    private List<Expectation> _expectations = new();
+    private readonly List<Expectation> _expectations = new();
 
     private MockLogsCollector(ITestOutputHelper output, string host = "localhost")
     {
@@ -57,9 +57,7 @@ public class MockLogsCollector : IDisposable
     {
         var collector = new MockLogsCollector(output, host);
 
-        var healhtzEndpoint = $"http://{(host == "*" ? "localhost" : host)}:{collector.Port}/healthz";
-
-        var healthzResult = await HealthzHelper.TestHealtzAsync(healhtzEndpoint, nameof(MockLogsCollector), output).ConfigureAwait(false);
+        var healthzResult = await collector._listener.VerifyHealthzAsync();
 
         if (!healthzResult)
         {
@@ -170,12 +168,6 @@ public class MockLogsCollector : IDisposable
 
     private void HandleHttpRequests(HttpListenerContext ctx)
     {
-        if (ctx.Request.RawUrl.EndsWith("/healthz", StringComparison.OrdinalIgnoreCase))
-        {
-            CreateHealthResponse(ctx);
-            return;
-        }
-
         if (ctx.Request.RawUrl.Equals("/v1/logs", StringComparison.OrdinalIgnoreCase))
         {
             var logsMessage = ExportLogsServiceRequest.Parser.ParseFrom(ctx.Request.InputStream);
@@ -212,16 +204,6 @@ public class MockLogsCollector : IDisposable
 
         // We received an unsupported request
         ctx.Response.StatusCode = (int)HttpStatusCode.NotImplemented;
-        ctx.Response.Close();
-    }
-
-    private void CreateHealthResponse(HttpListenerContext ctx)
-    {
-        ctx.Response.ContentType = "text/plain";
-        var buffer = Encoding.UTF8.GetBytes("OK");
-        ctx.Response.ContentLength64 = buffer.LongLength;
-        ctx.Response.OutputStream.Write(buffer, 0, buffer.Length);
-        ctx.Response.StatusCode = (int)HttpStatusCode.OK;
         ctx.Response.Close();
     }
