+++ test/IntegrationTests/Helpers/MockMetricsCollector.cs
@@ -71,14 +71,12 @@ public class MockMetricsCollector : IDisposable
     {
         var collector = new MockMetricsCollector(output, host);
 
-        var healhtzEndpoint = $"http://{(host == "*" ? "localhost" : host)}:{collector.Port}/healthz";
-
-        var healthzResult = await HealthzHelper.TestHealtzAsync(healhtzEndpoint, nameof(MockMetricsCollector), output).ConfigureAwait(false);
+        var healthzResult = await collector._listener.VerifyHealthzAsync();
 
         if (!healthzResult)
         {
             collector.Dispose();
-            throw new InvalidOperationException($"Cannot start {nameof(MockMetricsCollector)}!");
+            throw new InvalidOperationException($"Cannot start {nameof(MockLogsCollector)}!");
         }
 
         return collector;
@@ -144,12 +142,6 @@ public class MockMetricsCollector : IDisposable
     {
         OnRequestReceived(ctx);
 
-        if (ctx.Request.RawUrl.EndsWith("/healthz", StringComparison.OrdinalIgnoreCase))
-        {
-            CreateHealthResponse(ctx);
-            return;
-        }
-
         if (ctx.Request.RawUrl.Equals("/v1/metrics", StringComparison.OrdinalIgnoreCase))
         {
             if (ShouldDeserializeMetrics)
@@ -177,16 +169,6 @@ public class MockMetricsCollector : IDisposable
 
         // We received an unsupported request
         ctx.Response.StatusCode = (int)HttpStatusCode.NotImplemented;
-        ctx.Response.Close();
-    }
-
-    private void CreateHealthResponse(HttpListenerContext ctx)
-    {
-        ctx.Response.ContentType = "text/plain";
-        var buffer = Encoding.UTF8.GetBytes("OK");
-        ctx.Response.ContentLength64 = buffer.LongLength;
-        ctx.Response.OutputStream.Write(buffer, 0, buffer.Length);
-        ctx.Response.StatusCode = (int)HttpStatusCode.OK;
         ctx.Response.Close();
     }
